feat(cardano): index executed scripts under a script_redeemers dimension
The scripts redeemers scan derives candidates from side-effect
dimensions, and governance redeemers leave no side effect any dimension
tags. Resolve the redeemer-to-script matching once, at index time, and
tag each executed script under a new script_redeemers dimension.

- the matcher moves to pallas_extras with a generic error type and
  gains vote and propose arms (voter credential, proposal policy
  script); the endpoints keep their Blockfrost-parity output, so the
  new purposes surface only as tags for now
- index_block tags every resolvable execution, phase-2-failed txs
  included: tags are candidates, the Blockfrost rule stays a query-time
  filter
- the scripts redeemers union gains the new dimension as its first key;
  stores synced before this commit lack the tags, and the union keeps
  those complete
- redb3 (deprecated backend) gets the table and arms so the dimension
  works there too
- the stele goldens re-pin: the canonical indexes layer carries one
  record per dimension, so its diffId, record count and inscription
  digest change deliberately
